fix: size limit

+++ classes/Anel.cpp
@@ -8,7 +8,7 @@ Anel::Anel() {
     this->efeito = "";
 }
 
-Anel::Anel(int tipo, std::string nome, Tempo duracao, float preco, int quantidade, bool equipado, int buff, int debuff, std::string efeito) 
+Anel::Anel(int tipo, std::string nome, Tempo duracao, long double preco, long long quantidade, bool equipado, long long buff, long long debuff, std::string efeito) 
 : Item(tipo, nome, duracao, preco, quantidade, equipado) {
     this->buff=buff;
     this->debuff=debuff;
@@ -19,19 +19,19 @@ std::string Anel::getEfeito() {
     return this->efeito;
 }
 
-int Anel::getBuff() {
+long long Anel::getBuff() {
     return this->buff;
 }
 
-int Anel::getDeBuff() {
+long long Anel::getDeBuff() {
     return this->debuff;
 }
 
-void Anel::setBuff(int buff) {
+void Anel::setBuff(long long buff) {
     this->buff=buff;
 }
 
-void Anel::setDeBuff(int debuff) {
+void Anel::setDeBuff(long long debuff) {
     this->debuff=debuff;
 }
