+++ classes/Item.cpp
@@ -3,7 +3,7 @@
 
 Item::Item() {}
 
-Item::Item(int tipo, std::string nome, Tempo duracao, float preco, int quantidade, bool equipado){
+Item::Item(int tipo, std::string nome, Tempo duracao, long double preco, long long quantidade, bool equipado){
     this->tipo = tipo;
     this->nome = nome;
     this->duracao = duracao;
@@ -29,21 +29,21 @@ void Item::setNome(std::string& nome){
 Tempo Item::getDuracao(){
     return this->duracao;
 }
-void Item::setDuracao(int duracao){
+void Item::setDuracao(long long duracao){
     this->duracao.setTempo(duracao);
 }
 
-float Item::getPreco(){
+long double Item::getPreco(){
     return this->preco;
 }
-void Item::setPreco(float preco){
+void Item::setPreco(long double preco){
     this->preco = preco;
 }
 
-int Item::getQuantidade(){
+long long Item::getQuantidade(){
     return this->quantidade;
 }
-void Item::setQuantidade(int quantidade){
+void Item::setQuantidade(long long quantidade){
     this->quantidade = quantidade;
 }
