+++ classes/Pocao.cpp
@@ -6,7 +6,7 @@ Pocao::Pocao() {
     this->efeito = "";
 }
 
-Pocao::Pocao(int tipo, std::string nome, Tempo duracao, float preco, int quantidade, bool equipado, std::string efeito) 
+Pocao::Pocao(int tipo, std::string nome, Tempo duracao, long double preco, long long quantidade, bool equipado, std::string efeito) 
 : Item (tipo, nome, duracao, preco, quantidade, equipado){
     this->efeito = efeito;
 }
