+++ headers/Inventario.h
@@ -32,6 +32,6 @@ class Inventario {
 
         Item* readOne(int index);
         int countByType(int tipo);
-        float PrecoTotal();
+        long double PrecoTotal();
         void relatorio(); // Qtd de elementos totais, qtd elementos por classe, preco total     
 };
