+++ headers/Item.h
@@ -7,13 +7,13 @@ class Item {
         int tipo;
         std::string nome;
         Tempo duracao;
-        float preco;
-        int quantidade;
+        long long preco;
+        long long quantidade;
         bool equipado;
     public:
         Item();
        
-        Item (int tipo, std::string nome, Tempo duracao, float preco, int quantidade, bool equipado); 
+        Item (int tipo, std::string nome, Tempo duracao, long double preco, long long quantidade, bool equipado); 
 
         int getTipo();
 
@@ -21,9 +21,9 @@ class Item {
 
         Tempo getDuracao();
 
-        float getPreco();
+        long double getPreco();
 
-        int getQuantidade();
+        long long getQuantidade();
 
         bool getEquipado();
 
@@ -31,11 +31,11 @@ class Item {
 
         void setNome(std::string& nome);
 
-        void setDuracao(int duracao);
+        void setDuracao(long long duracao);
 
-        void setPreco(float preco);
+        void setPreco(long double preco);
 
-        void setQuantidade(int quantidade);
+        void setQuantidade(long long quantidade);
 
         void setEquipado(bool equipado);
